added buggy blueprints for DeflateLite

+++ OKcompress/src/main/java/OKcompress/domain/HuffmanNode.java
@@ -5,7 +5,7 @@ package OKcompress.domain;
  * Nodes used in HuffmanHeaps and Huffman trees.
  */
 public class HuffmanNode implements Comparable<HuffmanNode> {
-    public byte byteValue;
+    public int byteValue;
     public int weight;
     public HuffmanNode leftChild;
     public HuffmanNode rightChild;
@@ -14,7 +14,7 @@ public class HuffmanNode implements Comparable<HuffmanNode> {
     }
 
 
-    public HuffmanNode(byte byteValue, int weight) {
+    public HuffmanNode(int byteValue, int weight) {
         this.byteValue = byteValue;
         this.weight = weight;
     }
